creates TxForID component
